updated project to .net standard 2.0, and multi-targeting net 4.5 and .net 40 still. Didn't yet update the test/website projects
