+++ .gitignore
@@ -109,3 +109,4 @@ _UpgradeReport_Files/
 Backup*/
 UpgradeLog*.XML
 /.vs/config/applicationhost.config
+/.vs
