+++ Dapper.SimpleCRUD.sln
@@ -1,10 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.27004.2009
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dapper.SimpleCRUD", "Dapper.SimpleCRUD\Dapper.SimpleCRUD.csproj", "{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dapper.SimpleCRUDTests", "Dapper.SimpleCRUDTests\Dapper.SimpleCRUDTests.csproj", "{A4F4A06E-D179-4251-A232-AEF4CE9AD9B5}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{4C0A1156-5B42-4415-9306-39E265A360EC}"
@@ -14,9 +12,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{4C0A11
 		.nuget\NuGet.targets = .nuget\NuGet.targets
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dapper.SimpleCRUD NET45", "Dapper.SimpleCRUD NET45\Dapper.SimpleCRUD NET45.csproj", "{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DemoWebsite", "DemoWebsite\DemoWebsite.csproj", "{BCF52CCE-03CF-443E-BC02-7D92A1E78D60}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Dapper.SimpleCRUD", "Dapper.SimpleCRUD\Dapper.SimpleCRUD.csproj", "{9C42445F-4A30-43A1-A480-445B25CE2BD8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,16 +26,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Debug|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Debug|Mixed Platforms.Build.0 = Release|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{D88D1D75-0BBD-4E66-ABE0-A077C85DC302}.Release|x86.ActiveCfg = Release|Any CPU
 		{A4F4A06E-D179-4251-A232-AEF4CE9AD9B5}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{A4F4A06E-D179-4251-A232-AEF4CE9AD9B5}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
 		{A4F4A06E-D179-4251-A232-AEF4CE9AD9B5}.Debug|Mixed Platforms.Build.0 = Debug|x86
@@ -48,16 +36,6 @@ Global
 		{A4F4A06E-D179-4251-A232-AEF4CE9AD9B5}.Release|Mixed Platforms.Build.0 = Release|x86
 		{A4F4A06E-D179-4251-A232-AEF4CE9AD9B5}.Release|x86.ActiveCfg = Release|x86
 		{A4F4A06E-D179-4251-A232-AEF4CE9AD9B5}.Release|x86.Build.0 = Release|x86
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{A54F7AB7-F6DE-43AA-9B31-2065CC6F16D8}.Release|x86.ActiveCfg = Release|Any CPU
 		{BCF52CCE-03CF-443E-BC02-7D92A1E78D60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BCF52CCE-03CF-443E-BC02-7D92A1E78D60}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BCF52CCE-03CF-443E-BC02-7D92A1E78D60}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -68,8 +46,23 @@ Global
 		{BCF52CCE-03CF-443E-BC02-7D92A1E78D60}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{BCF52CCE-03CF-443E-BC02-7D92A1E78D60}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{BCF52CCE-03CF-443E-BC02-7D92A1E78D60}.Release|x86.ActiveCfg = Release|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Debug|x86.Build.0 = Debug|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Release|x86.ActiveCfg = Release|Any CPU
+		{9C42445F-4A30-43A1-A480-445B25CE2BD8}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6D7DEFD4-79ED-4D07-8527-FA62F312BCD4}
 	EndGlobalSection
 EndGlobal
